fixed the routing issues maybe? * model_controller_routing_path means nest/users * controller_routing_path means nest/session * model_controller_routing_name = users, (used for users_path etc and must be singularized for things like new_user_path) * controller_routing_name       = session (used for session_path, new_session_path, etc.) * controller_controller_name    = sessions, the name of the controller
If any of this is wrong, at least is should hopefully be easier to resolve with the nice overly-prolix semantic generator names.

I tried to make the names correct for nested resources.  However, the app explodes into horrible flames with anything like "generate Admin::User Session", and in fact with "generate users session".  I don\'t know when this broke, but I can't justify taking the time to fix it unless someone complains. Other than that, the specs and stories all pass for the three activation versions AFAICT.