Restore method recently_activated? on the model when using activation but not stateful or aasm.
If you run the generator with the --include-activation parameter but not either of --aasm or --stateful, then after adding the recommended observer to environment.rb, the generated specs start failing with: undefined method `recently_activated?' for #<User:0x7fa37d8beb38>

I tracked down the problem and it seems to be introduced near commit 0a29c1 or another one in that day (may 18, 2008).
This patch re-adds the method. But right now this method is triplicated on 3 files... not too DRY, but works and seems to keep modularization.